end of test1

+++ c_test/int/sha256.h
@@ -22,7 +22,7 @@ typedef unsigned int  WORD;             // 32-bit word, change to "long" for 16-
 typedef struct {
 	BYTE data[64];
 	WORD datalen;
-	unsigned int bitlen;
+	unsigned long long bitlen;
 	WORD state[8];
 } SHA256_CTX;
